feat: move vans pages
